expose colored/formatted consoleOutput in issues and summary of callback
